fix: change Nested File browser to a functional component
